Add the ability to produce keyed messages.
If the client presents a KeyedMessage, use the key when constructing the
message to sent to Kafka. This can be used to control to which partition
a message will be sent. (All messages with the same key will be sent to
the same partition and will therefore be consumed in the order in which
they were produced).